% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_RAMPMM.R
\name{PMM}
\alias{PMM}
\title{Function to perform Pattern Mixture Model for the Repeated Attempt Models (RAM-PMM)}
\usage{
PMM(Dataset, Modelfile, Response = "Y2", BaselineResponse = "Y1",
  Attempts = "R", Intervention = "Z", NumberofIndicators = 3,
  n.chains = 2, n.adapt = 1000, AttemptsMax = 3, ParamSens = 0)
}
\arguments{
\item{Dataset}{A data file ( txt/csv) having coulmns that are named and placed in a specific order.}

\item{Modelfile}{A JAGS model file that rjags will use to simulate a model via JAGS installed at User's system.}

\item{Response}{Column having values for 'Response'. Users must either have name 'Y2' for this column in their data or provide the name for this column as per their dataset.}

\item{BaselineResponse}{Column having values for 'Baseline Response'. Users must either have name 'Y1' for this column in their data or provide the name for this column as per their dataset.}

\item{Attempts}{Column having values 'Number of Attempts'. Users must either have name 'R' for this column in their data or provide the name for this column as per their dataset.}

\item{Intervention}{Column having values for 'Intervention'. Users must either have name 'Z' for this column in their data or provide the name for this column as per their dataset.}

\item{NumberofIndicators}{Number of Indicator Centers (Xs). The User dataset must have Xs starting from 5th column and ending as the last column of the dataset.}

\item{n.chains}{Parallel chains for the model}

\item{n.adapt}{The number of iterations for adaptation.}

\item{AttemptsMax}{Maximum Number of Attempts to collect the output}

\item{ParamSens}{The Sensitivity Parameter C; Possible Values 0,1,2,3.}
}
\value{
The summary of the posterior distribution. User will get an object named 'Results' in the global environment. Corresponding values for any particular column can be obtained by Results$columnname.
}
\description{
It runs the nonignorable analyses of the RAM-PMM approach on a dataset provided by the user
}

