% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projectDataDocumentation.R
\docType{data}
\name{projectData}
\alias{projectData}
\title{Dataset file for the RAMPMM function}
\format{A data frame with 409 rows and 7 variables:
\describe{
  \item{Z}{InterventionValues, bernoulli distribution}
  \item{Y2}{ResponseValues, Normal distribution}
  \item{Y1}{BaselineResponseValues, Normal distribution}
  \item{R}{NumberofAttemptsValues, Categorical distribution}
  \item{X1}{FirstCenterIndicators, Center Membership Dummy Variable}
  \item{X2}{SecondCenterIndicators, Center Membership Dummy Variable}
  \item{X3}{ThirdCenterIndicators, Center Membership Dummy Variable}
  ...
}}
\usage{
projectData
}
\description{
A dataset containing the columns as we wanted them
}
\keyword{datasets}

