% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAM_PMM.R
\name{RAMPMM}
\alias{RAMPMM}
\title{Function to perform Pattern Mixture Model for the Repeated Attempt Models (RAM-PMM)}
\usage{
RAMPMM(Dataset, Modelfile, NumberofIndicators = 7, n.chains = 2,
  n.adapt = 1000, Intervention = "Z", Response = "Y2",
  BaselineResponse = "Y1", Attempts = "R")
}
\arguments{
\item{Dataset}{A data file ( txt/csv) having coulmns named and placed in a fixed order.}

\item{Modelfile}{A WINBUGS model file that rjags will use to simulate a model via JAGS installed at User's system.}

\item{NumberofIndicators}{Number of Indicator Centers (Xs). The User dataset must have Xs starting from 5th column and ending as the last column of the dataset.}

\item{n.chains}{Parallel chains for the model}

\item{n.adapt}{iterations for adaptations}

\item{Intervention}{Column having values for 'Intervention'. Users must either have name 'Z' for this column in their data or provide the name for this column as per their dataset.}

\item{Response}{Column having values for 'Response'. Users must either have name 'Y2' for this column in their data or provide the name for this column as per their dataset.}

\item{BaselineResponse}{Column having values for 'Baseline Response'. Users must either have name 'Y1' for this column in their data or provide the name for this column as per their dataset.}

\item{Attempts}{Column having values 'Number of Attempts'. Users must either have name 'R' for this column in their data or provide the name for this column as per their dataset.}
}
\value{
The summary of the posterior distribution. User will get an object in the global environment  (results). Corresponding values for any particular column can be obtained by results$columnname.
}
\description{
It runs the nonignorable analyses of the RAM_PMM on the dataset provided by the user
}

