% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAM_PMM_fixed.R
\name{RAM_PMM_fixed}
\alias{RAM_PMM_fixed}
\title{Function to perform RAM_PMM}
\usage{
RAM_PMM_fixed(Dataset, Modelfile, NumberofIndicators)
}
\arguments{
\item{file}{having the data and model specifications}
}
\value{
returns the summary
}
\description{
It runs the nonignorable analyses for the RAM_PMM on the provided dataset
}

